requirements.txt
requests>=2.31.0
python-dateutil>=2.8.2
browser-use>=0.1.0
playwright>=1.40.0
openai>=1.0.0
python-dotenv>=1.0.0
